discord copie/requirements.txt
discord.py
python-dotenv
requests
mkdocs
mkdocs-material
mkdocs-material-extensions
pymdown-extensions